test(Picker): migrate test files to typescript
