+++ src/Picker/test/DropdownMenuGroupSpec.tsx
@@ -18,7 +18,10 @@ describe('picker - DropdownMenuGroup', () => {
       </DropdownMenuGroup>
     );
 
-    assert.equal(instance.querySelector('.rs-dropdown-menu-group-title').textContent, 'title');
+    assert.equal(
+      (instance.querySelector('.rs-dropdown-menu-group-title') as HTMLElement).textContent,
+      'title'
+    );
   });
 
   it('Should have a role', () => {
