+++ src/Picker/test/DropdownMenuItemSpec.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 import ReactTestUtils from 'react-dom/test-utils';
 
 import { getDOMNode } from '@test/testUtils';
-import DropdownMenuItem from '../DropdownMenuCheckItem';
+import DropdownMenuItem from '../DropdownMenuItem';
 
-describe('picker - DropdownMenuCheckItem', () => {
+describe('picker - DropdownMenuItem', () => {
   it('Should output a item', () => {
     const Title = 'Title';
     const instance = getDOMNode(<DropdownMenuItem title="title">{Title}</DropdownMenuItem>);
@@ -14,20 +14,21 @@ describe('picker - DropdownMenuCheckItem', () => {
   });
 
   it('Should be active', () => {
-    const instance = getDOMNode(<DropdownMenuItem title="title" active />);
+    const instance = getDOMNode(<DropdownMenuItem classPrefix="item" title="title" active />);
 
-    assert.include(instance.querySelector('.rs-checkbox').className, 'rs-checkbox-checked');
+    assert.include((instance.querySelector('.rs-item') as HTMLElement).className, 'item-active');
   });
 
   it('Should be disabled', () => {
-    const instance = getDOMNode(<DropdownMenuItem title="title" disabled />);
+    const instance = getDOMNode(<DropdownMenuItem classPrefix="item" title="title" disabled />);
 
-    assert.include(instance.querySelector('.rs-checkbox').className, 'rs-checkbox-disabled');
+    assert.include((instance.querySelector('.rs-item') as HTMLElement).className, 'item-disabled');
   });
 
   it('Should be focus', () => {
-    const instance = getDOMNode(<DropdownMenuItem title="title" focus />);
-    assert.include(instance.querySelector('.rs-checkbox').className, 'item-focus');
+    const instance = getDOMNode(<DropdownMenuItem classPrefix="item" title="title" focus />);
+
+    assert.include((instance.querySelector('.rs-item') as HTMLElement).className, 'item-focus');
   });
 
   it('Should call onSelect callback', done => {
@@ -36,7 +37,7 @@ describe('picker - DropdownMenuCheckItem', () => {
     };
     const instance = getDOMNode(<DropdownMenuItem title="title" onSelect={doneOp} />);
 
-    ReactTestUtils.Simulate.change(instance.querySelector('input'));
+    ReactTestUtils.Simulate.click(instance.querySelector('.rs-dropdown-menu-item') as HTMLElement);
   });
 
   it('Should call onKeyDown callback', done => {
@@ -45,30 +46,9 @@ describe('picker - DropdownMenuCheckItem', () => {
     };
     const instance = getDOMNode(<DropdownMenuItem title="title" onKeyDown={doneOp} />);
 
-    ReactTestUtils.Simulate.keyDown(instance.querySelector('input'));
-  });
-
-  it('Should call onCheck callback', () => {
-    const onSelectSpy = sinon.spy();
-    const onCheckeSpy = sinon.spy();
-
-    const instance = getDOMNode(
-      <DropdownMenuItem title="title" onCheck={onCheckeSpy} onSelectItem={onSelectSpy} />
+    ReactTestUtils.Simulate.keyDown(
+      instance.querySelector('.rs-dropdown-menu-item') as HTMLElement
     );
-
-    ReactTestUtils.Simulate.click(instance.querySelector('.rs-checkbox-wrapper'));
-
-    assert.ok(onCheckeSpy.calledOnce);
-    assert.ok(onSelectSpy.calledOnce);
-  });
-
-  it('Should call onSelectItem callback', done => {
-    const doneOp = () => {
-      done();
-    };
-    const instance = getDOMNode(<DropdownMenuItem title="title" onSelectItem={doneOp} />);
-
-    ReactTestUtils.Simulate.click(instance.querySelector('.rs-checkbox'));
   });
 
   it('Should have a role', () => {
@@ -85,5 +65,10 @@ describe('picker - DropdownMenuCheckItem', () => {
     const fontSize = '12px';
     const instance = getDOMNode(<DropdownMenuItem style={{ fontSize }} />);
     assert.equal(instance.style.fontSize, fontSize);
+  });
+
+  it('Should have a custom className prefix', () => {
+    const instance = getDOMNode(<DropdownMenuItem classPrefix="custom-prefix" />);
+    assert.ok((instance.querySelector('span') as HTMLElement).className.match(/\bcustom-prefix\b/));
   });
 });
