+++ src/Picker/test/PickerToggleSpec.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import ReactTestUtils from 'react-dom/test-utils';
 import { fireEvent, render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
+import sinon from 'sinon';
 
 import Toggle from '../PickerToggle';
 import { getDOMNode } from '@test/testUtils';
